+++ examples/resources/believe_ticket_sale/import.sh
@@ -1,0 +1,1 @@
+$ terraform import believe_ticket_sale.example '<ticket_sale_id>'