+++ internal/services/biscuit/data_source_schema.go
@@ -16,6 +16,7 @@ var _ datasource.DataSourceWithConfigValidators = (*BiscuitDataSource)(nil)
 
 func DataSourceSchema(ctx context.Context) schema.Schema {
 	return schema.Schema{
+		Description: "Interactive endpoints for motivation and guidance",
 		Attributes: map[string]schema.Attribute{
 			"biscuit_id": schema.StringAttribute{
 				Required: true,
@@ -37,7 +38,7 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 				Computed:    true,
 			},
 			"type": schema.StringAttribute{
-				Description: "Type of biscuit\nAvailable values: \"classic\", \"shortbread\", \"chocolate_chip\", \"oatmeal_raisin\".",
+				Description: "Type of biscuit\nAvailable values: \"classic\", \"shortbread\", \"chocolate_chip\", \"oatmeal_raisin\", \"snickerdoodle\", \"lemon_drizzle\".",
 				Computed:    true,
 				Validators: []validator.String{
 					stringvalidator.OneOfCaseInsensitive(
@@ -45,6 +46,8 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 						"shortbread",
 						"chocolate_chip",
 						"oatmeal_raisin",
+						"snickerdoodle",
+						"lemon_drizzle",
 					),
 				},
 			},