+++ internal/services/biscuit/list_data_source_schema.go
@@ -17,6 +17,7 @@ var _ datasource.DataSourceWithConfigValidators = (*BiscuitsDataSource)(nil)
 
 func ListDataSourceSchema(ctx context.Context) schema.Schema {
 	return schema.Schema{
+		Description: "Interactive endpoints for motivation and guidance",
 		Attributes: map[string]schema.Attribute{
 			"max_items": schema.Int64Attribute{
 				Description: "Max items to fetch, default: 1000",
@@ -48,7 +49,7 @@ func ListDataSourceSchema(ctx context.Context) schema.Schema {
 							Computed:    true,
 						},
 						"type": schema.StringAttribute{
-							Description: "Type of biscuit\nAvailable values: \"classic\", \"shortbread\", \"chocolate_chip\", \"oatmeal_raisin\".",
+							Description: "Type of biscuit\nAvailable values: \"classic\", \"shortbread\", \"chocolate_chip\", \"oatmeal_raisin\", \"snickerdoodle\", \"lemon_drizzle\".",
 							Computed:    true,
 							Validators: []validator.String{
 								stringvalidator.OneOfCaseInsensitive(
@@ -56,6 +57,8 @@ func ListDataSourceSchema(ctx context.Context) schema.Schema {
 									"shortbread",
 									"chocolate_chip",
 									"oatmeal_raisin",
+									"snickerdoodle",
+									"lemon_drizzle",
 								),
 							},
 						},