+++ internal/services/quote/data_source_schema.go
@@ -22,6 +22,7 @@ var _ datasource.DataSourceWithConfigValidators = (*QuoteDataSource)(nil)
 
 func DataSourceSchema(ctx context.Context) schema.Schema {
 	return schema.Schema{
+		Description: "Memorable quotes from the show",
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				Computed: true,
@@ -80,7 +81,7 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 				Computed:    true,
 			},
 			"theme": schema.StringAttribute{
-				Description: "Primary theme of the quote\nAvailable values: \"belief\", \"teamwork\", \"curiosity\", \"kindness\", \"resilience\", \"vulnerability\", \"growth\", \"humor\", \"wisdom\", \"leadership\", \"love\", \"forgiveness\", \"philosophy\", \"romance\", \"cultural-pride\", \"cultural-differences\", \"antagonism\", \"celebration\", \"identity\", \"isolation\", \"power\", \"sacrifice\", \"standards\", \"confidence\", \"conflict\", \"honesty\", \"integrity\".",
+				Description: "Primary theme of the quote\nAvailable values: \"belief\", \"teamwork\", \"curiosity\", \"kindness\", \"resilience\", \"vulnerability\", \"growth\", \"humor\", \"wisdom\", \"leadership\", \"love\", \"forgiveness\", \"philosophy\", \"romance\", \"cultural-pride\", \"cultural-differences\", \"antagonism\", \"celebration\", \"identity\", \"isolation\", \"power\", \"sacrifice\", \"standards\", \"confidence\", \"conflict\", \"honesty\", \"integrity\", \"intimidation\", \"ambition\", \"narcissism\", \"maturity\".",
 				Computed:    true,
 				Validators: []validator.String{
 					stringvalidator.OneOfCaseInsensitive(
@@ -111,6 +112,10 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 						"conflict",
 						"honesty",
 						"integrity",
+						"intimidation",
+						"ambition",
+						"narcissism",
+						"maturity",
 					),
 				},
 			},
@@ -154,6 +159,10 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 							"conflict",
 							"honesty",
 							"integrity",
+							"intimidation",
+							"ambition",
+							"narcissism",
+							"maturity",
 						),
 					),
 				},
@@ -195,7 +204,7 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 						},
 					},
 					"theme": schema.StringAttribute{
-						Description: "Filter by theme\nAvailable values: \"belief\", \"teamwork\", \"curiosity\", \"kindness\", \"resilience\", \"vulnerability\", \"growth\", \"humor\", \"wisdom\", \"leadership\", \"love\", \"forgiveness\", \"philosophy\", \"romance\", \"cultural-pride\", \"cultural-differences\", \"antagonism\", \"celebration\", \"identity\", \"isolation\", \"power\", \"sacrifice\", \"standards\", \"confidence\", \"conflict\", \"honesty\", \"integrity\".",
+						Description: "Filter by theme\nAvailable values: \"belief\", \"teamwork\", \"curiosity\", \"kindness\", \"resilience\", \"vulnerability\", \"growth\", \"humor\", \"wisdom\", \"leadership\", \"love\", \"forgiveness\", \"philosophy\", \"romance\", \"cultural-pride\", \"cultural-differences\", \"antagonism\", \"celebration\", \"identity\", \"isolation\", \"power\", \"sacrifice\", \"standards\", \"confidence\", \"conflict\", \"honesty\", \"integrity\", \"intimidation\", \"ambition\", \"narcissism\", \"maturity\".",
 						Optional:    true,
 						Validators: []validator.String{
 							stringvalidator.OneOfCaseInsensitive(
@@ -226,6 +235,10 @@ func DataSourceSchema(ctx context.Context) schema.Schema {
 								"conflict",
 								"honesty",
 								"integrity",
+								"intimidation",
+								"ambition",
+								"narcissism",
+								"maturity",
 							),
 						},
 					},